Finish desktop version of Natours website!
Since the last commit, work was done finishing the footer section, and then creating a Popup Modal that transitions on and off the page when the "Book Now" buttons are hit. Anchor targetting was also used to make buttons take the user to various sections of the page. Still need to make it mobile-responsive, though! :(